+++ CHANGELOG.md
@@ -6,6 +6,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.9.0] - 2026-07-24
+
 ### Added
 
 - **GUI: Genio IDE integration** — when the desktop app is launched from